+++ caqtus/gui/condetrol/device_configuration_editors/__init__.py
@@ -1,9 +1,13 @@
 from .configurations_editor import DeviceConfigurationsDialog
-from .device_configuration_editor import DeviceConfigurationEditor
+from .device_configuration_editor import (
+    DeviceConfigurationEditor,
+    FormDeviceConfigurationEditor,
+)
 from .sequencer_configuration_editor import SequencerConfigurationEditor
 
 __all__ = [
     "DeviceConfigurationEditor",
     "DeviceConfigurationsDialog",
     "SequencerConfigurationEditor",
+    "FormDeviceConfigurationEditor",
 ]